Flasky/requirements.txt
flask
werkzeug
Flask-PyMongo
Scrapy
requests
gevent
flask-uploads
flask-mongoengine
apscheduler